feat(custom-sw): moved sw registration to `registration.js` from `index.js`
