+++ packages/falcon-client/src/serviceWorker/registration.js
